feat(geofence): move geofence computation to separate module
