+++ RouteAnalyzer/output.json
@@ -22,6 +22,8 @@
       "latitude" : 45.110796,
       "longitude" : 7.66392
     },
+    "areaRadiusKm" : 5.0,
+    "waypointsOutsideCount" : 5,
     "waypoints" : [ {
       "timestamp" : 1741432513586,
       "latitude" : 45.10988,
@@ -42,7 +44,6 @@
       "timestamp" : 1741432513631,
       "latitude" : 45.11122,
       "longitude" : 7.66383
-    } ],
-    "waypointsOutsideCount" : 5
+    } ]
   }
 }
